agno-building-ai-agents/agno-knowledge/pyproject.toml
[project]
name = "agno-knowledge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.1.10",
    "chonkie[openai]>=1.4.0",
    "lancedb>=0.25.2",
    "openai>=2.6.0",
    "pandas>=2.3.3",
    "pypdf>=6.1.3",
    "python-dotenv>=1.1.1",
]
